Move diff_render.py into watch/ package
